norminette updates, threaded rendering seperated
