+++ Makefile
@@ -1,11 +1,13 @@
 NAME = miniRT
+NAME_THREAD = miniRT_threaded
 LIBFT = ./incl/libft/libft.a
 GNL = ./incl/get_next_line/gnl.a
 MLX = ./incl/mlx/libmlx.a
 INC = -I ./incl
 OBJ_DIR = obj/
+OBJ_B_DIR = obj_thread/
 CC = cc
-CFLAGS = -O3 -pthread -g $(INC) -Wall -Wextra -Werror -fsanitize=address,undefined
+CFLAGS = -O3 -pthread -g $(INC) -Wall -Wextra -Werror
 MFLAG = -L /usr/X11R6/lib -lXext -lX11 -lm
 RM = rm -f
 
@@ -44,7 +46,45 @@ SRC = srcs/parser/checks.c \
       srcs/movement_misc.c \
       srcs/main.c
 
+SRC_B = srcs/parser/checks.c \
+      srcs/parser/parse_amb.c \
+      srcs/parser/parse_cam.c \
+      srcs/parser/parse_cyl.c \
+      srcs/parser/parse_lig.c \
+      srcs/parser/parse_obj.c \
+      srcs/parser/parse_plane.c \
+      srcs/parser/parse_sphere.c \
+      srcs/parser/parse_utils.c \
+      srcs/parser/utils_p.c \
+      srcs/parser/parse.c \
+      srcs/cam_ray.c \
+      srcs/error.c \
+      srcs/free.c \
+      srcs/free2.c \
+      srcs/intersect.c \
+      srcs/matrix_ops.c \
+      srcs/matrix_ops2.c \
+      srcs/movement.c \
+      srcs/movement_utils.c \
+      srcs/object_selection.c \
+      srcs/random_gen.c \
+      srcs/render_thread.c \
+      srcs/utils_color.c \
+      srcs/utils_color2.c \
+      srcs/utils_vec.c \
+      srcs/utils.c \
+      srcs/utils_print.c \
+      srcs/utils_selection.c \
+      srcs/utils_vec2.c \
+      srcs/utils_atof.c \
+      srcs/movement_misc.c \
+	  srcs/render_modes.c \
+	  srcs/utils_render.c \
+      srcs/main.c
+
 OBJ = $(addprefix $(OBJ_DIR), $(SRC:.c=.o))
+
+OBJ_B = $(addprefix $(OBJ_B_DIR), $(SRC_B:.c=.o))
 
 HEADERS = incl/minirt.h \
 
@@ -68,8 +108,18 @@ $(OBJ_DIR)%.o: %.c $(HEADERS)
 	@mkdir -p $(@D)
 	@$(CC) $(CFLAGS) -c $< -o $@
 
+thread: $(NAME_THREAD)
+
+$(NAME_THREAD): $(OBJ_B) $(LIBFT) $(GNL) $(MLX)
+	@$(CC) $(CFLAGS) $(OBJ_B) $(LIBFT) $(GNL) $(MLX) -o $(NAME_THREAD) $(MFLAG)
+
+$(OBJ_B_DIR)%.o: %.c $(HEADERS)
+	@mkdir -p $(@D)
+	@$(CC) $(CFLAGS) -c $< -o $@
+
 clean:
 	@$(RM) -r $(OBJ_DIR)
+	@$(RM) -r $(OBJ_B_DIR)
 	@make clean -C ./incl/libft
 	@make clean -C ./incl/get_next_line
