+++ scenes/energy.rt
@@ -1,4 +1,4 @@
-A  0.1                           255,255,255
+A  0.0                           255,255,255
 
 L  -0.75,1.75,-1.5         1           255,255,255
 C  0,1,-12        0,-0.2,1        65
