+++ scenes/eval_opy.rt
@@ -1,6 +1,6 @@
-A  0.0                           255,255,255
+A  0.001                           255,255,255
 
-L  6,1.25,-1.25         0.75           255,255,255
+L  -3.75,5.75,-2.5         1           255,84,126
 C  0,0.5,-12        0,-0.1,1        65
 
 sp -5,-1,1.75          5         10,10,10
