+++ srcs/utils_vec.c
@@ -6,7 +6,7 @@
 /*   By: root <root@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/04 10:24:48 by root              #+#    #+#             */
-/*   Updated: 2024/12/04 10:24:48 by root             ###   ########.fr       */
+/*   Updated: 2024/12/08 12:46:51 by root             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,41 +14,22 @@
 
 t_vec	vec(double x, double y, double z, double w)
 {
-	t_vec	vec;
-
-	vec.x = x;
-	vec.y = y;
-	vec.z = z;
-	vec.w = w;
-	return (vec);
+	return ((t_vec){x, y, z, w});
 }
 
 t_vec	vec_add(t_vec a, t_vec b)
 {
-	t_vec	ret;
-
-	ret.x = a.x + b.x;
-	ret.y = a.y + b.y;
-	ret.z = a.z + b.z;
-	return (ret);
+	return ((t_vec){a.x + b.x, a.y + b.y, a.z + b.z, a.w});
 }
 
 t_vec	vec_sub(t_vec a, t_vec b)
 {
-	t_vec	ret;
-
-	ret.x = a.x - b.x;
-	ret.y = a.y - b.y;
-	ret.z = a.z - b.z;
-	return (ret);
+	return ((t_vec){a.x - b.x, a.y - b.y, a.z - b.z, a.w});
 }
 
-double	vec_dot(t_vec a, t_vec b)
+inline double	vec_dot(t_vec a, t_vec b)
 {
-	double	ret;
-
-	ret = (a.x * b.x) + (a.y * b.y) + (a.z * b.z);
-	return (ret);
+	return ((a.x * b.x) + (a.y * b.y) + (a.z * b.z));
 }
 
 t_vec	vec_multiply(t_vec a, t_vec b)
